+++ msr/raplcap-cpuid.h
@@ -52,6 +52,9 @@ extern "C" {
 #define CPUID_MODEL_TIGERLAKE_L       0x8C
 #define CPUID_MODEL_TIGERLAKE         0x8D
 
+#define CPUID_MODEL_ALDERLAKE         0x97
+#define CPUID_MODEL_ALDERLAKE_L       0x9A
+
 #define CPUID_MODEL_XEON_PHI_KNL      0x57
 #define CPUID_MODEL_XEON_PHI_KNM      0x85
 
@@ -69,6 +72,8 @@ extern "C" {
 #define CPUID_MODEL_ATOM_GOLDMONT_PLUS  0x7A // Gemini Lake
 
 #define CPUID_MODEL_ATOM_TREMONT_D      0x86 // Jacobsville
+#define CPUID_MODEL_ATOM_TREMONT        0x96 // Elkhart Lake
+#define CPUID_MODEL_ATOM_TREMONT_L      0x9C // Jasper Lake
 //----
 
 /**