+++ RELEASES.md
@@ -1,6 +1,6 @@
 # Release Notes
 
-## [Unreleased]
+## [v0.8.0] - 2022-05-13
 
 ### Added
 
@@ -245,7 +245,7 @@
 * Initial public release
 
 
-[Unreleased]: https://github.com/powercap/raplcap/compare/v0.7.0...HEAD
+[v0.8.0]: https://github.com/powercap/raplcap/compare/v0.7.0...v0.8.0
 [v0.7.0]: https://github.com/powercap/raplcap/compare/v0.6.0...v0.7.0
 [v0.6.0]: https://github.com/powercap/raplcap/compare/v0.5.0...v0.6.0
 [v0.5.0]: https://github.com/powercap/raplcap/compare/v0.4.0...v0.5.0